refact(main,lib): move “absolute path from file” to AbsPathBuf
This is really a constructor for `AbsPathBuf`, so let’s move it there.
